+++ urishell/__init__.py
@@ -1,3 +1,0 @@
-from urishell.routes import register
-
-__all__ = ["register"]